perfect the loop unrolling benchmark

+++ src/loops.cpp
@@ -2,20 +2,22 @@
 // Created by matrix on 2023/11/23.
 //
 
+#include <cstdint>
+
 #include "loops.h"
 
 namespace loops {
 
-    int SumWithLoopUnrolling(int n) {
-        int res = 0;
+    int64_t SumWithLoopUnrolling(int64_t n) {
+        int64_t res = 0;
         for (int i = 0; i < n; i += 4) {
             res += i + (i + 1) + (i + 2) + (i + 3);
         }
         return res;
     }
 
-    int SumWithoutLoopUnrolling(int n) {
-        int res = 0;
+    int64_t SumWithoutLoopUnrolling(int64_t n) {
+        int64_t res = 0;
         for (int i = 0; i < n; ++i) {
             res += i;
         }
